LayoutTests/platform/mac-bigsur/imported/w3c/web-platform-tests/reporting/cross-origin-report-no-credentials.https.sub-expected.txt

CONSOLE MESSAGE: Refused to load https://localhost:9443/reporting/resources/fail.png because it does not appear in the img-src directive of the Content Security Policy.
Blocked access to external URL https://www1.localhost:9443/reporting/resources/report.py?reportID=GENERATED_REPORT_ID


FAIL Reporting endpoints did not receive credentials. assert_unreached: A report of csp-violation from https://localhost:9443/reporting/cross-origin-report-no-credentials.https.sub.html is not found. Reached unreachable code
